Update CardboardEye.cs
Avoid to see the temp RT when the RT is not create Finish in one Frame! I slove it in our project by add the two line work and I donot want to add they again when I upgrade the CardboardSDK,So please verify it and It's my pleasure it works!